move vtt_volt_data out of microsoft directory
